Temporary fix for un-sent SV replies.  Changes behavior for response to Change Address.  Add some method description comments.
A temporary fix for SV reply message until basic LocoNet message
handling is re-factored.  SV reply now uses the LocoNetClass::send()
method which loops until the message is sent, unless the maximum number
of retries is reached.  Also now attempts to send a LONG_ACK message if
the reply was not successfully sent.