Patch from Stefan Bormann:
There was a bug found in the central buffer module ln_buf.c of the EmbeddedLoconet project. Bodo Mertins, a fellow FREMO club member found and fixed this bug.

This module is in use in Alex' Modelrailroading with Arduino library, too. Alex, you probably want to apply the fix there soon.

I comitted the fix to the Subversion repository of EmbeddedLoconet.
The fix can be seen in this diff:
https://sourceforge.net/p/embeddedloconet/svn/933/tree//old-cvs/trunk/loconet/ln_buf.c?diff=484

The bug only has effect, when packets with uneven byte count are used on the LocoNet. When the opcode of a variable length packet is received into the very last byte of the buffer and the next byte of the packet, which is the length count of the packet, wraps to the very first byte of the buffer. In this rare situation, the decision which set of memcpy() calls are to be done is wrong and the wrong copies are done, which screws up memory content.

FREMO is using this library a lot in the FREDI handheld throttle and never saw this bug, because for train operation the only variable length packets - write and read slot - are even length. Lucky us...

Once you know the bug, it is easy to reproduce:

1) find out how big the buffer is in your device (let's assume the default of 128 bytes)
2) reset the device (e.g. power cycle)
3) send messages to the net that add up to 127 bytes, e.g. from PC
4) send a variable length message to the device that has an observable effect
5) nothing is observed