No longer ignore setSpeed(1) calls sometimes
- When making a normal (non-emergency) stop, the user requests speed 1,
but the value actually stored & sent to LocoNet is 0. If the user
happens to setSpeed(1) after another call to setFunction(),
setDirection(), or setSpeed() without first calling
process100msActions(), which is common when changing multiple things in
sequence, then the speed change will be deferred. It will set
myDeferredSpeed to 0. However, the next time that process100msActions()
executes, it will not send the necessary speed packet, because
myDeferredSpeed is zero! Thus, the requested speed change will not take
effect. It won't take effect 60 seconds later when the slot refresh
occurs either, because myDeferredSpeed==0 means it will use mySpeed
instead.
- To fix this, changed myDeferredSpeed from uint8_t to int16_t so that
we can use -1 as a special value indicating that no change is needed,
without conflicting with any valid speed values.